Improvements to handling deploy step
Added script to upload tarballs to S3 bucket. The original script is
adjusted such that it can work with both AWS S3 and other S3 bucket
hosting servers such as Minio. The script can also add metadata about
the PR for which the built artefact is uploaded. This could be used to
update status information during the ingestion procedure.

The configration settings have been optimized to work with upload
script.

The deploy implementation has been updated for the changed interface to
the upload script. Also, the file 'uploaded.txt' is now created/updated
when a file has been uploaded to an S3 bucket.